+++ generateDefaultData.js
@@ -1,18 +1,18 @@
 import generateData from './modules/generator'
 
 const defaultBrowserSupport = {
-  chrome: 46,
-  android: 4,
-  firefox: 40,
-  ios_saf: 8,
-  safari: 8,
+  chrome: 55,
+  android: 5,
+  firefox: 52,
+  ios_saf: 9,
+  safari: 9,
   ie: 11,
   ie_mob: 11,
   edge: 12,
-  opera: 16,
+  opera: 30,
   op_mini: 12,
-  and_uc: 9,
-  and_chr: 46
+  and_uc: 11,
+  and_chr: 55
 }
 
 generateData(defaultBrowserSupport, {