+++ modules/static/plugins/sizing.js
@@ -1,5 +1,5 @@
 /* @flow */
-const prefixes = ['-webkit-', '-moz-', '']
+const prefixes = ["-webkit-", "-moz-", ""];
 
 const properties = {
   maxHeight: true,
@@ -9,17 +9,17 @@ const properties = {
   columnWidth: true,
   minWidth: true,
   minHeight: true
-}
+};
 const values = {
-  'min-content': true,
-  'max-content': true,
-  'fill-available': true,
-  'fit-content': true,
-  'contain-floats': true
-}
+  "min-content": true,
+  "max-content": true,
+  "fill-available": true,
+  "fit-content": true,
+  "contain-floats": true
+};
 
 export default function sizing(property: string, value: any): ?Array<any> {
   if (properties.hasOwnProperty(property) && values.hasOwnProperty(value)) {
-    return prefixes.map(prefix => prefix + value)
+    return prefixes.map(prefix => prefix + value);
   }
 }